requirements.txt
click>=6.7
python-dateutil>=2.7.3
lxml>=4.2.3
requests>=2.19.1
